% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{calc_gamma}
\alias{calc_gamma}
\title{Calculate gamma}
\usage{
calc_gamma(se, fe, sc, fc, a, b, mu, sigma2, norm, c2)
}
\arguments{
\item{se, fe}{number of successes and failures on CYC}

\item{sc, fc}{number of successes and failures on MMF}

\item{a, b}{parameters of prior distribution of pC}

\item{mu, sigma2}{parameters of prior distribution of theta}

\item{norm}{normalising constant of joint posterior distribution of (pC, pE)}

\item{c2}{non-inferiority margin for the trial}
}
\value{
P{pE > pC - c2|data}
}
\description{
Calculate gamma
}
