% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{calc_pi}
\alias{calc_pi}
\title{Calculate pi}
\usage{
calc_pi(se, mu, sigma2, pmean, pvar, a, b, ne, sc, fc)
}
\arguments{
\item{se}{number of successes observed on MMF in the dataset}

\item{mu, sigma2}{parameters of prior distribution of theta}

\item{pmean, pvar}{posterior mean and variance of theta.}

\item{a, b}{parameters of prior distribution of pC}

\item{ne, sc, fc}{number randomised to MMF; number of successes and failures on CYC.}
}
\value{
\deqn{P{pE > pC|data}}
}
\description{
Calculate pi
}
