% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_funcs.R
\name{calc_thetavar}
\alias{calc_thetavar}
\title{Search for value of 1/sigma such that P{pE - pC > -margin} = gamma as elicited by 1 - Q4}
\usage{
calc_thetavar(sigmainv, mu_sigma, a, b, gamma, margin)
}
\arguments{
\item{sigmainv}{1/sigma where sigma is the hypothesized prior sd of theta}

\item{mu_sigma}{mu/sigma, ratio of the prior mean and sd of theta}

\item{a, b}{parameters of prior distribution of pC}

\item{gamma}{value of P{pE - pC > -margin} elicited form the expert (=1-q4)}

\item{margin}{non-inferiority margin for the trial}
}
\value{
P(pE - pC > -margin; 1/sigma) - gamma.
}
\description{
Using Simpsons integration to integrate over the joint prior distribution of (pE, pC).
}
