% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_funcs.R
\name{check_shape_pE_U_or_L}
\alias{check_shape_pE_U_or_L}
\title{check_shape_pE_U_or_L}
\usage{
check_shape_pE_U_or_L(int, gride)
}
\arguments{
\item{int}{}

\item{gride}{}
}
\description{
Checking to see whether marginal prior density for pE is a U or L shaped function of pE
Integrate prior pE density over the interval \link{lim1, lim2}.
If probability in this interval is less than would be the case under a flat density, we conclude
the density could be U or L shaped.
}
