% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_funcs.R
\name{logoddspc}
\alias{logoddspc}
\title{Calculate variance of prior distribution of log\link{pc/(1-pc)}}
\usage{
logoddspc(a, b)
}
\arguments{
\item{a, b}{parameters of prior distribution of pC}
}
\value{
prior variance of log\link{pC/(1-pC)} (used for evaluating prior ESS of log-odds)
}
\description{
Calculate variance of prior distribution of log\link{pc/(1-pc)}
}
