% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{pc_dens}
\alias{pc_dens}
\title{Posterior density of pC}
\usage{
pc_dens(gridc, fe, se, fc, sc, priorParm, norm)
}
\arguments{
\item{gridc}{}

\item{fe}{}

\item{se}{}

\item{fc}{}

\item{sc}{}

\item{priorParm}{}

\item{norm}{}
}
\description{
Posterior density of pC
}
