% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{pe_dens}
\alias{pe_dens}
\title{Posterior density of pE}
\usage{
pe_dens(gride, fe, se, fc, sc, priorParm, norm)
}
\arguments{
\item{gride}{vector of values of pE}

\item{fe, se, fc, sc}{number of successes and failures on MMF and CYC}

\item{priorParm}{output of priorcall() summarising elicited prior distributions}

\item{norm}{normalising constant of joint posterior distribution g(pC, pE|data)}
}
\value{
posterior density of pE eveluated at gride
}
\description{
Posterior density of pE
}
