% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{posterior_summaries}
\alias{posterior_summaries}
\title{Summaries of posterior distributions of pC, pE and theta given hypothetical dataset}
\usage{
posterior_summaries(n_mmf, mmf_succ, n_cyc, cyc_succ, priorParm, posterior40)
}
\arguments{
\item{n_mmf, mmf_succ}{(nE, SE): number of patients randomised to MMF and number of observed successes on experimental arm}

\item{n_cyc, cyc_succ}{(nC, SC): number of patients randomised to CYC and number of observed success on CYC}

\item{priorParm}{vector outputted by priorcall() containing summaries of prior distributions}

\item{posterior40}{logical variable indicating whether wish to calculate posterior distribution assuming nE+nC = 40
(if false assume wish to calculate posterior distribution assuming nE + nC = 20)}
}
\description{
x\link{1} = E(pC|data), x\link{2} = mode(pC|data), x\link{3} = SD(pC|data), (x\link{4}, x\link{5}) = 90\% posterior credibility interval for pC, x\link{18} = normalising constant of g(pC, pE|data)
x\link{6} = E(pE|data), x\link{7} = mode(pE|data), x\link{8} = SD(pE|data), (x\link{9}, x\link{10}) = 90\% posterior credibility interval for pE, x\link{19} = normalising constant of g(pC, pE|data)
x\link{11} = E(theta|data), x\link{12} = mode(theta|data), x\link{13} = SD(theta|data), (x\link{14}, x\link{15}) = 90\% posterior credibility interval for theta,
x\link{16} = P{pE > pC|data}, x\link{17} = P{pE - pC > -margin|data},  x\link{20} = normalising constant of joint posterior distribution f(theta, pC|data)
}
