% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{postSumry}
\alias{postSumry}
\title{Summaries of posterior distributions given hypothetical dataset}
\usage{
postSumry(n_mmf, mmf_succ, n_cyc, cyc_succ, priorParm, posterior40)
}
\arguments{
\item{n_mmf}{}

\item{mmf_succ}{}

\item{n_cyc}{}

\item{cyc_succ}{}

\item{priorParm}{}

\item{posterior40}{}
}
\description{
Summaries of posterior distributions given hypothetical dataset
}
