% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_funcs.R
\name{prior_e}
\alias{prior_e}
\title{evaluate the marginal prior distribution of pE.}
\usage{
prior_e(a, b, mu, sigma2)
}
\arguments{
\item{a, b}{parameters of prior distribution of pC}

\item{mu, sigma2}{parameters of prior distribution of theta}
}
\value{
dataframe containing prior expectation, mode, sd,
limits of 90\% credibility interval and 25th percentile.
}
\description{
evaluate the marginal prior distribution of pE.
}
