% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{prior_summaries}
\alias{prior_summaries}
\title{Calculate summaries of the prior distributions of pC, pE and theta}
\usage{
prior_summaries(q1, q2, q3, q4, expert = "", out_dir = "plots")
}
\arguments{
\item{q1, q2, q3, q4}{Expert's answers to the Day 1 elicitation questions}

\item{expert}{Character string giving the expert's initials}

\item{out_dir}{Output folder}
}
\value{
A vector summarising properties of the prior distributions of pC, pE and theta
x\link{1}, x\link{2} parameters of beta prior distribution for pC, x\link{3} = ESS of log(pC/(1-pC)) given elicited prior for pC
x\link{4} = E(pE), x\link{5} = mode(pE), x\link{6} = SD(pE), (x\link{7}, x\link{8}) = 90\% credibility interval, x\link{14} = quantile such that P(pE <= x\link{14}) = 0.25
x\link{9} = E(theta), x\link{10} = var(theta), x\link{11} = P(pE > pC), x\link{12} = P(pE < pC - 0.1), x\link{13} = ESS of theta

Also outputs a file saved as "expert-priorplot.pdf" in the current working directory containing plots of the prior densities
and a file "expert-D1answer.txt" storing the expert's answers to the elicitation questions.
}
\description{
q1 is the answer to Day 1 elicitation question (i) mode(pc)
q2 is the answer to Day 1 elicitation question (ii) eliciting q s.t. P(pC > q) = 0.75
q3 is the answer to Day 1 elicitation question (iii) eliciting P(theta > 0)
q4 is the answer to Day 1 elicitation question (iv) eliciting P(pE - pC < -margin)
}
