% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{priorcall}
\alias{priorcall}
\title{priorcall}
\usage{
priorcall(q1, q2, q3, q4, expert = "", out_dir = "plots")
}
\arguments{
\item{q1}{}

\item{q2}{}

\item{q3}{}

\item{q4}{}

\item{expert}{}
}
\description{
priorcall
}
