% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_pC}
\alias{prob_pC}
\title{prob_pC}
\usage{
prob_pC(q, prob, norm, se, sc, fe, fc, a, b, mu, sigma2)
}
\arguments{
\item{q}{hypothesized 100*prob percentile of the distribution.    iu}

\item{prob}{indicator (posterior =1 implies inferences concern posterior distribution)}

\item{norm}{normalisation constant of joint posterior distribution of (pC, pE)}

\item{se, sc, fe, fc}{number of successes and failures on treatments CYC and MMF}

\item{a, b}{parameters of priors pC~Beta(a,b)}

\item{mu, sigma2}{parameters of priors theta~N(mu, sigma2)}
}
\value{
P{pC <= q|data} - prob.
}
\description{
prob_pC
}
