% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_theta}
\alias{prob_theta}
\title{Title}
\usage{
prob_theta(q, prob, norm, se, sc, fe, fc, a, b, mu, sigma2, postexp, postsd)
}
\arguments{
\item{q}{}

\item{prob}{}

\item{norm}{}

\item{se, sc, fe, fc}{}

\item{a, b}{}

\item{mu, sigma2}{}

\item{postexp, postsd}{}
}
\description{
Title
}
