% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{theta_dens}
\alias{theta_dens}
\title{Evaluate posterior density of theta}
\usage{
theta_dens(gridt, fe, se, fc, sc, priorParm, norm)
}
\arguments{
\item{gridt}{vector of values of theta}

\item{fe, se, fc, sc}{number of successes and failures on MMF and CYC}

\item{priorParm}{output of priorcall() summarising elicited prior distributions}

\item{norm}{normalising constant of joint posterior distribution g(pC, theta|data)}
}
\value{
posterior density of theta evaluated at gridt
}
\description{
Evaluate posterior density of theta
}
