% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_funcs.R
\name{thetavar}
\alias{thetavar}
\title{Posterior variance of theta}
\usage{
thetavar(a, b, sc, se, fc, fe, mu, sigma2)
}
\arguments{
\item{a, b}{}

\item{sc, se, fc, fe}{}

\item{mu, sigma2}{}
}
\description{
Posterior variance of theta
}
