% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{dist_plot_data}
\alias{dist_plot_data}
\title{Prior and posterior densities of pC, pE and theta ready for plotting}
\usage{
dist_plot_data(
  n_mmf,
  mmf_succ,
  n_cyc,
  cyc_succ,
  postParm,
  priorParm,
  parmInd,
  postind
)
}
\arguments{
\item{n_mmf}{}

\item{mmf_succ}{}

\item{n_cyc}{}

\item{cyc_succ}{}

\item{postParm}{}

\item{priorParm}{}

\item{parmInd}{}

\item{postind}{}
}
\description{
Prior and posterior densities of pC, pE and theta ready for plotting
}
