% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_prior.R
\name{ess_pc}
\alias{ess_pc}
\title{Effective sample size (ESS) pc}
\usage{
ess_pc(a, b)
}
\arguments{
\item{a, b}{parameters of the prior distribution of \deqn{pC~Beta(a,b)}}
}
\value{
Effective sample size of prior for \deqn{\log[pC/(1-pc)]}
(expressed in terms of effective observations on control arm)
}
\description{
Calculate prior effective sample size of \deqn{log[pc/(1-pC)]}.
}
