% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{pc_dens}
\alias{pc_dens}
\title{Posterior density of pC}
\usage{
pc_dens(gridc, fe, se, fc, sc, priorParm, norm)
}
\arguments{
\item{gridc}{vector of values of pC}

\item{fe, se, fc, sc}{number of successes and failures on MMF and CYC}

\item{priorParm}{output of priorcall() summarising elicited prior distributions}

\item{norm}{normalising constant of joint posterior distribution g(pC, pE|data)}
}
\value{
posterior density of pC eveluated at gridc
}
\description{
Posterior density of pC
}
