% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{pe_dens}
\alias{pe_dens}
\title{Posterior density of pE}
\usage{
pe_dens(gride, fe, se, fc, sc, priorParm, norm)
}
\arguments{
\item{gride}{}

\item{fe, se, fc, sc}{number of successes and failures on MMF and CYC}

\item{priorParm}{}

\item{norm}{}
}
\description{
Posterior density of pE
}
