% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_funcs.R
\name{post_theta}
\alias{post_theta}
\title{Posterior theta}
\usage{
post_theta(a, b, sc, se, fc, fe, mu, sigma2, c2)
}
\arguments{
\item{a, b}{}

\item{sc, se, fc, fe}{}

\item{mu, sigma2}{}

\item{c2}{}
}
\value{
dataframe containing posterior expectation, SD, mode,
limits of the 90\% credibility interval of theta and the normalisation
constant of the posterior joint distribution of (pC, theta)
}
\description{
Posterior theta
}
