% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{posterior_summaries}
\alias{posterior_summaries}
\title{Summaries of posterior distributions given hypothetical dataset}
\usage{
posterior_summaries(n_mmf, mmf_succ, n_cyc, cyc_succ, priorParm, posterior40)
}
\arguments{
\item{n_mmf}{}

\item{mmf_succ}{}

\item{n_cyc}{}

\item{cyc_succ}{}

\item{priorParm}{}

\item{posterior40}{is posterior 40 selected? Logical}
}
\description{
Summaries of posterior distributions given hypothetical dataset
}
