% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_funcs.R
\name{prior_beta}
\alias{prior_beta}
\title{Find the parameters of prior Beta distribution of pc~Beta(a,b)}
\usage{
prior_beta(mode, percentile25)
}
\arguments{
\item{mode}{prior mode of pc as identified by Day 1 elicitation Q1}

\item{percentile25}{25th percentile of prior distribution for pC as identified by Day 1 elicitation Q2.}
}
\value{
vector (a,b) (parameters of Beta prior distribution for pC).
}
\description{
Find the parameters of prior Beta distribution of pc~Beta(a,b)
}
