% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_pC}
\alias{prob_pC}
\title{Title}
\usage{
prob_pC(q, prob, norm, se, sc, fe, fc, a, b, mu, sigma2)
}
\arguments{
\item{q}{}

\item{prob}{}

\item{norm}{}

\item{se, sc, fe, fc}{}

\item{a, b}{}

\item{mu, sigma2}{}
}
\description{
Title
}
