% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_pE}
\alias{prob_pE}
\title{Probability pE}
\usage{
prob_pE(q, prob, norm, a, b, mu, sigma2, sc, fc, se, fe, posterior)
}
\arguments{
\item{q}{hypothesized 100*prob percentile of the distribution.}

\item{prob}{Probability}

\item{norm}{normalising constant of joint posterior distribution of (pE, pC) given data}

\item{a, b, mu, sigma2}{parameters of priors pC~Beta(a,b), theta~N(mu, sigma2)}

\item{sc, fc, s, fe}{number of successes and failures on treatments CYC and MMF}

\item{posterior}{indicator (posterior =1 implies inferences concern posterior distribution)}
}
\value{
P{pE <= q} - prob.
}
\description{
Functions for calculating P{pE <= q}, P{pE <= q|data}, P{pC <= q|data}, P{theta <= q|data}.
}
