% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_theta}
\alias{prob_theta}
\title{prob_theta}
\usage{
prob_theta(q, prob, norm, se, sc, fe, fc, a, b, mu, sigma2, postexp, postsd)
}
\arguments{
\item{q}{hypothesized 100*prob percentile of the distribution.}

\item{prob}{}

\item{norm}{normalisation constant of joint posterior distribution of (pC, theta)}

\item{se, sc, fe, fc}{number of successes and failures on treatments CYC and experimental arm}

\item{a, b}{parameters of priors pC~Beta(a,b)}

\item{mu, sigma2}{parameters of priors theta~N(mu, sigma2)}

\item{postexp, postsd}{expectation and standard deviation of posterior distribution of theta.}
}
\value{
P{theta <= q|data} - prob.
}
\description{
prob_theta
}
