% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_funcs.R
\name{thetavar}
\alias{thetavar}
\title{produce a course estimate of posterior variance of theta}
\usage{
thetavar(a, b, sc, se, fc, fe, mu, sigma2)
}
\arguments{
\item{a, b}{parameters of prior distribution of pC}

\item{sc, se, fc, fe}{number of successes and failures observed on CYC and experimental arm}

\item{mu, sigma2}{parameters of prior distribution of theta}
}
\value{
estimate of variance of the posterior distribution of theta.
}
\description{
produce a course estimate of posterior variance of theta
}
